requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
baidu-aip==4.16.10
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
edge-tts==6.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.4
paho-mqtt==1.6.1
playsound==1.2.2
pvporcupine==2.1.0
PyAudio==0.2.13
pydub==0.25.1
pygame==2.3.0
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
